+++ src/imio/smartweb/core/browser/forms.py
@@ -6,11 +6,18 @@ from plone.dexterity.browser.add import DefaultAddView
 from plone.dexterity.browser.edit import DefaultEditForm
 from plone.z3cform import layout
 from z3c.form.interfaces import DISPLAY_MODE
+from z3c.form.interfaces import HIDDEN_MODE
 
 
 class CustomAddForm(DefaultAddForm):
     css_class = "tabbed-form-with-toggle"
     enable_form_tabbing = False
+
+    def updateFields(self):
+        super(CustomAddForm, self).updateFields()
+        if "ILeadImageBehavior.image_caption" in self.fields:
+            # We don't use leadimage caption anywhere
+            self.fields["ILeadImageBehavior.image_caption"].mode = HIDDEN_MODE
 
 
 class CustomAddView(DefaultAddView):
@@ -22,8 +29,10 @@ class CustomEditForm(DefaultEditForm):
     enable_form_tabbing = False
 
     def updateFields(self):
-        super(DefaultEditForm, self).updateFields()
-
+        super(CustomEditForm, self).updateFields()
+        if "ILeadImageBehavior.image_caption" in self.fields:
+            # We don't use leadimage caption anywhere
+            self.fields["ILeadImageBehavior.image_caption"].mode = HIDDEN_MODE
         if IDefaultPages.providedBy(self.context):
             # We need to make exclude_from_nav field read only as it is used
             # for default pages to exclude them from sitemap / navigation / ...