+++ src/imio/smartweb/core/utils.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+
+from collective.taxonomy.interfaces import ITaxonomy
+from plone import api
+from Products.CMFPlone.utils import base_hasattr
+from zope.component import getSiteManager
+
+
+def get_category(context):
+    if not base_hasattr(context, "category_name"):
+        return
+    field_name = "taxonomy_{}".format(context.category_name)
+    taxonomy_name = "collective.taxonomy.{}".format(context.category_name)
+    term = getattr(context, field_name)
+    if not term:
+        return
+    current_lang = api.portal.get_current_language()[:2]
+    sm = getSiteManager()
+    utility = sm.queryUtility(ITaxonomy, name=taxonomy_name)
+    value = utility.translate(
+        term,
+        context=context,
+        target_language=current_lang,
+    )
+    return value